fixed duplication
